refactor(daemon): extract health snapshot + status derivation into supervisor-health.ts; plug untrack gap in watch-fault early-return
